Rethemed app and Pagination
Add working pagination buttons at the top of the page that render based
on if the server tells the client if it is the last or first page of
results. Removed box shadows and size differences in favor of dynamic
sizing of city items and consistent theme with color pallette.