Project refactor
Refactored project structure by domain. Added a common package that
stores mediators and other objects that operate between domains. Added
User in tandem with ApplicationUser. User is for business logic relating
to the domain. ApplicationUser only considers security. Added
CommodityExchange object that mediates between user, city, and commodity
and allows a user to supply a city with a commodity given that they have
it. More refactors to come and still very prototypical.