Added UserLoader
Changed List<Commodity> to Map<Commodity, Integer> in User to allow for
the concept of a Commodity supply versus an infinite amount of
commodities. Can only give the commodity to a city if the user has more
than 0 of the commodity. Added interface to UserLoaderImpl as the Spring
container will not correctly inject the UserLoader dependency unless it
has its own interface. Added UserLoader which gets all of the
commodities and application users, finds each User mapping of the
application user and randomly adds a configurable amount of commodities
to the user.